+++ failover-lookup/src/main/java/com/societegenerale/failover/lookup/BeanFactoryFailoverExpiryExtractor.java
@@ -1,0 +1,62 @@
+/*
+ * Copyright 2022-2026, Société Générale All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.societegenerale.failover.lookup;
+
+import com.societegenerale.failover.core.expiry.AbstractFailoverExpiryExtractor;
+import org.jspecify.annotations.NonNull;
+import org.springframework.beans.BeansException;
+import org.springframework.beans.factory.BeanFactory;
+import org.springframework.beans.factory.BeanFactoryAware;
+import org.springframework.beans.factory.config.ConfigurableBeanFactory;
+
+import java.time.temporal.ChronoUnit;
+
+import static java.util.Objects.requireNonNull;
+
+/**
+ * {@link AbstractFailoverExpiryExtractor} that resolves expiry expressions via
+ * Spring's {@link ConfigurableBeanFactory#resolveEmbeddedValue}, supporting
+ * property placeholders (e.g. {@code ${my.ttl}}) and SpEL expressions.
+ *
+ * @author Anand Manissery
+ */
+public class BeanFactoryFailoverExpiryExtractor extends AbstractFailoverExpiryExtractor implements BeanFactoryAware {
+
+    private ConfigurableBeanFactory beanFactory;
+
+    /**
+     * Injects the Spring {@link BeanFactory} used for expression resolution.
+     *
+     * @param beanFactory the bean factory; must implement {@link ConfigurableBeanFactory}
+     * @throws BeansException if setting the bean factory fails
+     */
+    @Override
+    public void setBeanFactory(@NonNull BeanFactory beanFactory) throws BeansException {
+        this.beanFactory = (ConfigurableBeanFactory) beanFactory;
+    }
+
+    @Override
+    protected long resolveExpiryDuration(String expression) {
+        return Long.parseLong(requireNonNull(beanFactory.resolveEmbeddedValue(expression)));
+    }
+
+    @Override
+    protected ChronoUnit resolveExpiryUnit(String expression) {
+        return ChronoUnit.valueOf(requireNonNull(beanFactory.resolveEmbeddedValue(expression)).toUpperCase());
+    }
+
+}