+++ failover-store-multitenant/src/test/java/com/societegenerale/failover/store/multitenant/MultiTenantInmemoryIntegrationTest.java
@@ -18,7 +18,7 @@ package com.societegenerale.failover.store.multitenant;
 
 import com.societegenerale.failover.core.payload.ReferentialPayload;
 import com.societegenerale.failover.core.store.FailoverStore;
-import com.societegenerale.failover.store.FailoverStoreInmemory;
+import com.societegenerale.failover.store.inmemory.FailoverStoreInmemory;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;